Move the cache logic in CoreMethodNodeManager to its own class
